+++ packages/blocks/src/components/slash-menu/config.ts
@@ -30,7 +30,7 @@ import {
 } from '../../__internal__/utils/index.js';
 import { clearMarksOnDiscontinuousInput } from '../../__internal__/utils/virgo.js';
 import { copyBlock } from '../../page-block/default/utils.js';
-import { formatConfig } from '../../page-block/utils/const.js';
+import { formatConfig } from '../../page-block/utils/format-config.js';
 import {
   onModelTextUpdated,
   updateBlockType,
