+++ packages/blocks/src/page-block/utils/const.ts
@@ -1,32 +1,17 @@
 import {
-  BoldIcon,
   CopyIcon,
   DatabaseTableViewIcon,
-  InlineCodeIcon,
-  ItalicIcon,
-  LinkIcon,
   SHORT_KEY,
-  StrikethroughIcon,
-  UnderlineIcon,
 } from '@blocksuite/global/config';
-import {
-  assertExists,
-  type BaseBlockModel,
-  type Page,
-} from '@blocksuite/store';
+import { assertExists, type Page } from '@blocksuite/store';
 
 import { copyBlocks } from '../../__internal__/clipboard/utils.js';
-import { createLink } from '../../__internal__/rich-text/link-node/index.js';
 import type { AffineTextAttributes } from '../../__internal__/rich-text/virgo/types.js';
 import { showDatabaseModal } from '../../components/database-modal/index.js';
 import { toast } from '../../components/toast.js';
-import {
-  getCurrentCombinedFormat,
-  handleFormat,
-} from '../../page-block/utils/index.js';
 import { getCurrentBlockRange } from '../../std.js';
 
-type ActionProps = {
+export type ActionProps = {
   page: Page;
   abortController?: AbortController;
   /**
@@ -34,84 +19,6 @@ type ActionProps = {
    */
   format?: AffineTextAttributes;
 };
-
-export const formatConfig = [
-  {
-    id: 'bold',
-    name: 'Bold',
-    icon: BoldIcon,
-    hotkey: `${SHORT_KEY}+b`,
-    activeWhen: (format: AffineTextAttributes) => 'bold' in format,
-    showWhen: (models: BaseBlockModel[]) => noneCodeBlockSelected(models),
-    action: ({ page }: ActionProps) => {
-      handleFormat(page, 'bold');
-    },
-  },
-  {
-    id: 'italic',
-    name: 'Italic',
-    icon: ItalicIcon,
-    hotkey: `${SHORT_KEY}+i`,
-    activeWhen: (format: AffineTextAttributes) => 'italic' in format,
-    showWhen: (models: BaseBlockModel[]) => noneCodeBlockSelected(models),
-    action: ({ page }: ActionProps) => {
-      handleFormat(page, 'italic');
-    },
-  },
-  {
-    id: 'underline',
-    name: 'Underline',
-    icon: UnderlineIcon,
-    hotkey: `${SHORT_KEY}+u`,
-    activeWhen: (format: AffineTextAttributes) => 'underline' in format,
-    showWhen: (models: BaseBlockModel[]) => noneCodeBlockSelected(models),
-    action: ({ page }: ActionProps) => {
-      handleFormat(page, 'underline');
-    },
-  },
-  {
-    id: 'strike',
-    name: 'Strikethrough',
-    icon: StrikethroughIcon,
-    hotkey: `${SHORT_KEY}+shift+s`,
-    activeWhen: (format: AffineTextAttributes) => 'strike' in format,
-    showWhen: (models: BaseBlockModel[]) => noneCodeBlockSelected(models),
-    action: ({ page }: ActionProps) => {
-      handleFormat(page, 'strike');
-    },
-  },
-  {
-    id: 'code',
-    name: 'Code',
-    icon: InlineCodeIcon,
-    hotkey: `${SHORT_KEY}+e`,
-    activeWhen: (format: AffineTextAttributes) => 'code' in format,
-    showWhen: (models: BaseBlockModel[]) => noneCodeBlockSelected(models),
-    action: ({ page }: ActionProps) => {
-      handleFormat(page, 'code');
-    },
-  },
-  {
-    id: 'link',
-    name: 'Link',
-    icon: LinkIcon,
-    hotkey: `${SHORT_KEY}+k`,
-    activeWhen: (format: AffineTextAttributes) => 'link' in format,
-    // Only can show link button when selection is in one line paragraph
-    showWhen: (models: BaseBlockModel[]) =>
-      models.length === 1 &&
-      noneCodeBlockSelected(models) &&
-      // can't create link when selection includes reference node
-      // XXX get loose format at here is not a good practice
-      !getCurrentCombinedFormat(models[0].page, true).reference,
-    action: ({ page, abortController, format }: ActionProps) => {
-      createLink(page);
-      if (format && abortController && !('link' in format)) {
-        abortController.abort();
-      }
-    },
-  },
-];
 
 const DATABASE_WHITE_LIST = ['affine:list', 'affine:paragraph'];
 
@@ -157,7 +64,3 @@ export const actionConfig = [
     },
   },
 ];
-
-function noneCodeBlockSelected(models: BaseBlockModel[]) {
-  return !models.every(model => model.flavour === 'affine:code');
-}
